pyproject.toml
[tool.poetry]
name = "pygptcourse"
version = "0.1.2"
description = "Face detection and tracking using USB T-Shirt launcher and rpi camera module to demonstrate how ChatGPT can speed up development and writing clean and SOLID code."
authors = ["G Shiva"]
license = "Apache"
readme = "README.md"
packages = [{ include = "pygptcourse", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
opencv-python = "4.5.5.62"
face-recognition = "1.3.0"
pyusb = "^1.2.1"
setuptools = "^68.2.2"
opencv-contrib-python = "4.5.5.62"
python-dotenv = "^1.0.0"
opentelemetry-api = "^1.22.0"
opentelemetry-sdk = "^1.22.0"
opentelemetry-exporter-otlp = "^1.22.0"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
flake8 = "^6.0.0"
pytest = "^7.3.2"
ipython = "^8.14.0"
mypy = "^1.7.1"
pre-commit = "^3.6.0"
coverage = "^7.3.4"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "piwheels.org"
url = "https://www.piwheels.org/simple"
priority = "supplemental"

[[tool.poetry.source]]
name = "PyPI"
priority = "primary"

[tool.pytest.ini_options]
filterwarnings = ["ignore:pkg_resources is deprecated::face_recognition_models"]

[tool.coverage.run]
omit = [".*", "*/tests/*", "*/acceptance_test/*", "*/site-packages/*"]

[tool.coverage.report]
fail_under = 47
